debug: fix WARN_ON_ONCE() for modules
Mike Galbraith reported a situation where a WARN_ON_ONCE in DRM code
turned into an oops.  As it turns out, WARN_ON_ONCE seems to be
completely broken when called from a module.

The bug was introduced with the following commit:

  19d436268dde ("debug: Add _ONCE() logic to report_bug()")

That commit changed WARN_ON_ONCE() to move its 'once' logic into the bug
trap handler.  It requires a writable bug table so that the BUGFLAG_DONE
bit can be written to the flags to indicate the first warning has
occurred.

The bug table was made writable for vmlinux, which relies on
vmlinux.lds.S and vmlinux.lds.h for laying out the sections.  However,
it was not made writable for modules, which rely on the elf section
flags.

Reported-by: Mike Galbraith <efault@gmx.de>
Fixes: 19d436268dde ("debug: Add _ONCE() logic to report_bug()")
Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>