x86/asm: Fix inline asm call constraints for clang
For inline asm statements which have a CALL instruction, we list the
stack pointer as a constraint to convince GCC to ensure the frame
pointer is set up first:

  static inline void foo()
  {
  	register void *__sp asm(_ASM_SP);
  	asm("call bar" : "+r" (__sp))
  }

Unfortunately, that pattern causes clang to corrupt the stack pointer.

The fix is easy: convert the stack pointer register variable to a global
variable.

It should be noted that the end result is different based on the GCC
version.  With GCC 6.4, this patch has exactly the same result as
before:

	defconfig	defconfig-nofp	distro		distro-nofp
before	9820389		9491555		8816046		8516940
after	9820389		9491555		8816046		8516940

With GCC 7.2, however, GCC's behavior has changed because of the
conversion of the register variable to a global.  That somehow convinces
GCC 7.2 to *always* set up the frame pointer before inserting *any*
inline asm.  (Therefore, listing the variable as an output constraint is
a no-op and is no longer necessary.)  It's a bit overkill, but the
performance impact should be negligible.  And in fact, there's a nice
improvement with frame pointers disabled:

	defconfig	defconfig-nofp	distro		distro-nofp
before	9796316		9468236		9076191		8790305
after	9796957		9464267		9076381		8785949

So in summary, while listing the stack pointer as an output constraint
is no longer necessary for newer versions of GCC, it's still needed for
older versions.

Reported-by: Matthias Kaehlcke <mka@chromium.org>
Suggested-by: Andrey Ryabinin <aryabinin@virtuozzo.com>
Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>