x86/paravirt: Introduce pv_alternative patching
With CONFIG_PARAVIRT enabled, the kernel .text is littered with a bunch
of calls to pv_irq_ops function pointers, like:

  callq  *0xffffffff81e3a400 (pv_irq_ops.save_fl)

In non-Xen paravirt environments -- including native, KVM, Hyper-V, and
VMware -- the above code gets patched by native_patch() to look like
this instead:

   pushfq
   pop    %rax
   nopl   0x0(%rax,%rax,1)

So in most scenarios, there's a mismatch between vmlinux and the actual
runtime code.  This mismatch hurts debuggability and makes the assembly
code harder to understand.

It also causes the ORC unwinder to produce KASAN warnings like:

  BUG: KASAN: stack-out-of-bounds in deref_stack_reg+0x123/0x140

This warning is due to the fact that objtool doesn't know about
parainstructions, so it doesn't know about the "pushfq; pop %rax"
sequence above.

Fix both of these issues (debuggability and ORC KASAN warnings) by using
the alternatives infrastucture to put the native instructions in .text
by default.  Then, when booting from a hypervisor, replace the native
instructions with pv ops calls.

The pv ops calls need to be set up much earlier than when alternatives
are normally applied.  So put these alternatives in a dedicated
".pv_alternatives" section.

This has a downside that, in some cases, this makes the paravirt
patching even more complicated than before.  For example, when running
on KVM, the code is first patched from native to "call
*pv_irq_ops.save_fl" by apply_pv_alternatives().  But then, later, it's
patched back to the native code again, by apply_paravirt().

But overall I think it will cause less confusion because the vmlinux
code now represents the actual runtime state of the code in the most
common paravirt cases (native, KVM, Hyper-V, and VMware).

Reported-by: Sasha Levin <alexander.levin@verizon.com>
Suggested-by: Andy Lutomirski <luto@kernel.org>
Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>