x86: Don't include asm/x86_init.h in asm/setup.h
This is a preparation to allow using 'struct timespec64' in asm/x86_init.h.

Unfortunately, using a forward declaration for timespec64 is not possible
since it is defined as a macro on 64-bit architectures, and including
linux/time64.h breaks compilation of arch/x86/boot/, which runs in realmode
and can't use many of the regular kernel headers.

As a workaround, stop including asm/x86_init.h. This works fine for the
realmode boot code since it does not require any of the x86_init.h
contents, and setup.h doesn't either.

However, a couple of other files need x86_init.h and rely on it being
included indirectly, so add an explicit include in there now.

[ tglx: Massaged changelog to 'imperative mood' ]

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Cc: y2038@lists.linaro.org
Cc: Joerg Roedel <joro@8bytes.org>
Cc: iommu@lists.linux-foundation.org
Link: https://lkml.kernel.org/r/20171019105451.2892046-1-arnd@arndb.de