x86/dumpstack: Fix partial register dumps
The show_regs_safe() logic is wrong.  When there's an iret frame, it
prints the entire pt_regs (most of which is random stack data) instead
of just the five registers at the end.

Also, show_regs_safe() is poorly named: the on_stack() checks aren't for
safety.  Rename the function to show_regs_if_on_stack() and add a
comment to explain why the checks are needed.

Tested-by: Alexander Tsoy <alexander@tsoy.me>
Fixes: b02fcf9ba121 ("x86/unwinder: Handle stack overflows more gracefully")
Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>