x86/paravirt: Remove 'noreplace-paravirt' cmdline option
The 'noreplace-paravirt' option disables paravirt patching, leaving the
original pv indirect calls in place.

That's very incompatible with retpolines, unless we want to uglify
paravirt even further and convert the indirect calls to retpolines.

As far as I can tell, the option doesn't seem to be useful for much
other than introducing surprising corner cases and leaving the kernel
vulnerable to Spectre v2.  So just remove it.

Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>