ftrace: Use '__used' instead of '__noclone' for selftest target functions
Objtool is reporting the following warnings with CONFIG_KCOV:

  kernel/trace/trace_selftest_dynamic.o: warning: objtool: trace_selftest_dynamic_test_func()+0x5: call without frame pointer save/setup
  kernel/trace/trace_selftest_dynamic.o: warning: objtool: trace_selftest_dynamic_test_func2()+0x5: call without frame pointer save/setup

The warnings are correct: the functions are indeed missing the frame
pointer setup.  This is apparently caused by the '__noclone' macro's use
of the '__optimize__("no-tracer") attribute', combined with the
'-fsanitize-coverage=trace-pc' flag.

Instead of 'noinline noclone', what we actually want is 'noinline used'.
That tells GCC that the symbol is used elsewhere and should not be
removed nor inlined.

Thanks to Martin Jambor for explaining these function attributes in more
detail.

Fixes: dd3dad0d716d ("ftrace: Mark function tracer test functions noinline/noclone")
Reported-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Josh Poimboeuf <jpoimboe@redhat.com>