KVM: VMX: Let the compiler save/load RDX during vCPU-run
Per commit c20363006af6 ("KVM: VMX: Let gcc to choose which registers
to save (x86_64)"), the only reason RDX is saved/loaded to/from the
stack is because it was specified as an input, i.e. couldn't be marked
as clobbered (ignoring the fact that "saving" it to a dummy output
would indirectly mark it as clobbered).

Now that RDX is no longer an input, mark it as clobbered and zero it
out to prevent speculative use.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>