KVM: VMX: Use #defines in place of immediates in VM-Enter inline asm
Using #defines is the last step moving the inline asm to a proper asm
sub-routine.  Eliminating the immediates allows a nearly verbatim move,
e.g. quotes, newlines, tabs and __stringify need to be dropped, but
other than those cosmetic changes the only function change is to replace
the final "jmp" with a "ret".

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>