KVM: VMX: Use RAX as the scratch register during vCPU-run
Prepare for making __vmx_vcpu_run() callable from C code.  That means
returning the result in RAX.  Since RAX will be used to return the
result, use it as the scratch register as well to make the code readable
and to document that the scratch register is more more less arbitrary.

Signed-off-by: Sean Christopherson <sean.j.christopherson@intel.com>